pyproject.toml
[project]
name = "study-duo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "confluent-kafka>=2.13.0",
    "fastapi[standard]>=0.133.1",
    "langchain-classic>=1.0.1",
    "langchain-core>=1.2.16",
    "langchain-huggingface>=1.2.0",
    "mcp-use>=1.6.0",
]
